tic_tac_toe_images/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==20.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
